pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "deepgram-sdk>=4.2.0",
    "elevenlabs>=2.3.0",
    "groq>=0.26.0",
    "livekit>=1.0.8",
    "livekit-agents>=1.0.23",
    "livekit-plugins-openai>=1.0.23",
    "livekit-plugins-silero>=1.0.23",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "websockets>=15.0.1",
]
